security/manager/ssl/tests/mochitest/stricttransportsecurity/test_sts_privatebrowsing.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->

<!DOCTYPE HTML>
<html>
<head>
  <title>opens additional content that should be converted to https</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

  <script class="testbody" type="text/javascript">
  SimpleTest.waitForExplicitFinish();

  const STSPATH = "/tests/security/ssl/stricttransportsecurity";

  const NUM_TEST_FRAMES = 4;
  var testframes = {
    'samedom':
      {'url':     "http://example.com" + STSPATH + "/verify.sjs",
        'expected': {'plain': 'SECURE',
                     'subdom': 'SECURE',
                     'nosts': 'INSECURE'}},
    'subdom':
      {'url':     "http://test1.example.com" + STSPATH + "/verify.sjs",
        'expected': {'plain': 'INSECURE',
                     'subdom': 'SECURE',
                     'nosts': 'INSECURE'}},
    'otherdom':
      {'url':     "http://example.org" + STSPATH + "/verify.sjs",
        'expected': {'plain': 'INSECURE',
                     'subdom': 'INSECURE',
                     'nosts': 'INSECURE'}},
    'alreadysecure':
      {'url':     "https://test2.example.com" + STSPATH + "/verify.sjs",
        'expected': {'plain': 'SECURE',
                     'subdom': 'SECURE',
                     'nosts': 'SECURE'}},
  };

  // This is how many sub-tests (testframes) in each round.
  // When the round begins, this will be initialized.
  var testsleftinround = 0;
  var currentround = "";

  var gPBSvc = null;
  var gPrefSvc = null;

  function _getServices() {
    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

    // not all apps will have the private browsing service.
    try {
      gPBSvc = Components.classes["@mozilla.org/privatebrowsing;1"]
                         .getService(Components.interfaces.nsIPrivateBrowsingService);
    } catch (ex) {
      SimpleTest.todo(false, "PB service is not available, will skip dependent tests");
    }

    // Not all apps will have the preference service.
    try {
      gPrefSvc = Components.classes["@mozilla.org/preferences-service;1"]
                           .getService(Components.interfaces.nsIPrefService)
                           .QueryInterface(Components.interfaces.nsIPrefBranch);
    } catch (ex) {
      SimpleTest.todo(false, "Pref service is not available, won't be able to manage preferences");
    }
  }
  _getServices();

  function startRound(round) {
    currentround = round;
    testsleftinround = NUM_TEST_FRAMES;
    SimpleTest.info("TESTS LEFT IN ROUND " + currentround + ": " + testsleftinround);
    var frame = document.createElement("iframe");
    frame.setAttribute('id', 'ifr_bootstrap');
    frame.setAttribute('src', "https://example.com" + STSPATH +
                              "/" + round + "_bootstrap.html");
    document.body.appendChild(frame);
  }

  function loadVerifyFrames(round) {
    for (var test in testframes) {
      var frame = document.createElement("iframe");
      frame.setAttribute('id', 'ifr_' + test);
      frame.setAttribute('src', testframes[test].url + '?id=' + test);
      document.body.appendChild(frame);
    }
  }

  /* Messages received are in this format:
   *  (BOOTSTRAP|SECURE|INSECURE) testid
   * For example: "BOOTSTRAP subdom"
   *          or: "INSECURE otherdom"
   */
  function onMessageReceived(event) {

    // otherwise, it's a test result
    var result = event.data.split(/\s+/);
    if (result.length != 2) {
      SimpleTest.ok(false, event.data);
      return;
    }

    if (result[0] === "BOOTSTRAP") {
      loadVerifyFrames(currentround);
      return;
    }

    // check if the result (SECURE/INSECURE) is expected for this round/test
    // combo
    dump_STSState();
    SimpleTest.is(result[0], testframes[result[1]].expected[currentround],
                             "in ROUND " + currentround +
                             ", test " + result[1]);
    testsleftinround--;

    // if this round is complete...
    if (testsleftinround < 1) {
      SimpleTest.info("DONE WITH ROUND " + currentround);
      // remove all the iframes in the document
      document.body.removeChild(document.getElementById('ifr_bootstrap'));
      for (var test in testframes)
        document.body.removeChild(document.getElementById('ifr_' + test));
      currentround = "";

      // And advance to the next test.
      // Defer this so it doesn't muck with the stack too much.
      SimpleTest.executeSoon(nextTest);
    }
  }

  function test_sts_before_private_mode() {
    dump_STSState();
    SimpleTest.info("Not in private browsing mode");
    startRound('plain');
  }

  function test_sts_in_private_mode() {
    dump_STSState();
    SimpleTest.info("Entering private browsing mode ...");
    netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
    if (gPrefSvc)
      gPrefSvc.setBoolPref("browser.privatebrowsing.keep_current_session", true);
    gPBSvc.privateBrowsingEnabled = true;
    SimpleTest.info("... done");
    dump_STSState();
    startRound('subdom');
  }

  function test_sts_after_exiting_private_mode() {
    dump_STSState();
    SimpleTest.info("Exiting private browsing mode ...");
    netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
    gPBSvc.privateBrowsingEnabled = false;
    if (gPrefSvc)
      gPrefSvc.clearUserPref("browser.privatebrowsing.keep_current_session");
    SimpleTest.info("... done");
    dump_STSState();
    startRound('nosts');
  }

  function clean_up_sts_state() {
    // erase all signs that this test ran.
    SimpleTest.info("Cleaning up STS data");
    netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
    const Cc = Components.classes;
    const Ci = Components.interfaces;
    var ios = Cc["@mozilla.org/network/io-service;1"]
                .getService(Ci.nsIIOService);
    var thehost = ios.newURI("http://example.com", null, null);
    var stss = Cc["@mozilla.org/stsservice;1"]
                 .getService(Ci.nsIStrictTransportSecurityService);
    stss.removeStsState(thehost);
    dump_STSState();
    SimpleTest.executeSoon(nextTest);
  }

function dump_STSState() {
    netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
    var stss = Components.classes["@mozilla.org/stsservice;1"]
          .getService(Components.interfaces.nsIStrictTransportSecurityService);
    SimpleTest.info("State of example.com: " + stss.isStsHost("example.com"));
}

  // these are executed in the order presented.
  // 0.  test that STS works before entering private browsing mode.
  //     (load sts-bootstrapped "plain" tests)
  //  ... clear any STS data ...
  var tests = [
    test_sts_before_private_mode,
    clean_up_sts_state
  ];
  // 1.  test that STS works in private browsing mode
  //     (load sts-bootstrapped "subdomain" tests)
  // 2.  test that after exiting private browsing, STS data is forgotten
  //     (verified with non-sts-bootstrapped pages)
  //  ... clear any STS data ...
  // Skip these tests if there is no private mode support.
  if (gPBSvc) {
    tests = tests.concat(
      test_sts_in_private_mode,
      test_sts_after_exiting_private_mode,
      clean_up_sts_state
    );
  }

  function nextTest() {
    SimpleTest.executeSoon(tests.length ? tests.shift() : SimpleTest.finish);
  }

  // listen for calls back from the sts-setting iframe and then
  // the verification frames.
  window.addEventListener("message", onMessageReceived, false);
  window.addEventListener('load', nextTest, false);
  </script>
</head>

<body>
  This test will load some iframes and do some tests.

</body>
</html>
